requirements.txt
fastapi
uvicorn[standard]
pypdf
langchain-text-splitters
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
PyMuPDF
Pillow
rank-bm25
pydantic
numpy